+++ NEWS.md
@@ -1,3 +1,4 @@
+* First version that works with full SEAPSUT workflow.
 * Added automated spell checking to package checks.
 * Added new target `report_output_folder` to plan and tests.