+++ NEWS.md
@@ -1,3 +1,10 @@
+# SEAPSUTWorkflow 0.1.6 (2020-08-26)
+
+* First version that works with full SEAPSUT workflow.
+* Added automated spell checking to package checks.
+* Added new target `reports_output_folder` to plan and tests.
+
+
 # SEAPSUTWorkflow 0.1.5 (2020-08-19)
 
 * Fixed bugs with completion code when put into production.